feat: add `useColorMode` composable
