+++ src/runtime/vue2/composables.ts
@@ -1,0 +1,6 @@
+import { useNuxtApp } from '#app'
+import type { ColorModeInstance } from '../types'
+
+export const useColorMode = () => {
+  return useNuxtApp().$colorMode as ColorModeInstance
+}
