+++ src/runtime/vue3/composables.ts
@@ -1,0 +1,6 @@
+import { useState } from '#app'
+import type { ColorModeInstance } from '../types'
+
+export const useColorMode = () => {
+  return useState('color-mode').value as ColorModeInstance
+}
